Move ComposerTextareaIcons to TextareaIcons
